Use structs for payload types
When receiving data in handlers, the previous practice was to accept a
`Document` as JSON and then pull out the elements you cared about. This
has two issues, firstly it makes it difficult to see what a route
expects at a glance, and secondly it allows part of the route to process
before failing with a 422.

Instead, this uses structs to determine what each route should expect.
This means that `actix-web` will never call a handler with an unfinished
payload, allowing it to fail earlier on a request. It also allows
developers to see the arguments for a function more easily based on
their struct definitions.