Check status codes after protocol requests
When the DCL queries the API server for information about a client, such
as whether they should be authorised, check the response code of the
request for error messages. If the request was not successful, return an
error from the protocol and do not add the client to the nodepool.

Fixes #269.