Refactor the API server (alongside small changes elsewhere) (#271)
* Allow `build_server` to return the server

Separate the library logic from the binary by allowing the library to
return a server and the binary to actually start it.

* Change `model_performance` method

Change the `get_model_performance` method to use a `GET` request instead
of a `POST` request for better semantics, updating the route itself, its
registration within the `api-server` and the request method from the
frontend.

* Swap to `fern` for logging and minimise output

Follow in the footsteps of the other components and use `fern` for
logging. As the `api-server` logs `actix-*` components as well as
itself, add a method to create a logger with different levels for
modules other than the crate itself.

Minimise the output of the logging to be a little cleaner. For the
moment, we don't care where requests come from and we aren't interested
in their User Agents, so just log the response code and the first line
of the request, which includes the method and route.

* Simplify state management

Change the state of each request to store a `mongodb::Database` instead
of a `mongodb::Client`. This removes a line in each request handler and
allows the database name to be changed more easily.

* Remove the `db_name` parameter of `State`

Remove the database name from the `State` struct as it wasn't being used
and was hardcoded to `sybl` anyway.

* Remove the `sleep`s from test setup

When creating each of the job performance instances, avoid sleeping
inbetween their constructors and instead modify the time slightly
afterwards.

* Remove some unwraps from the `api-server`

Continue removing `unwrap` calls from the `api-server`, finalising this
by having it contain no calls at all in the route-handling. This means
all errors should be propagated back to the requester, easing debugging
if strange things begin happening.

* Refactor the error type to `ServerError`

Rename the error type to `ServerError` and move it to `src/error.rs`.
Add some type aliases for it, to allow for using `ServerResult<T>` when
returning a given type and also handling errors and `ServerResponse` for
the general `Result<HttpResponse, ServerError>`.

This shortens some of the routes a little as `rustfmt` doesn't need to
wrap lines as often.

* Clean up import statements

Use a more consistent methodology for import statements to make it
slightly easier to read.

* Standardise on `crypto::{clean, clean_json}`

Standardise usage of `crypto::{clean, clean_json}` by using the full
path instead of importing it for clarity.

* Use structs for payload types

When receiving data in handlers, the previous practice was to accept a
`Document` as JSON and then pull out the elements you cared about. This
has two issues, firstly it makes it difficult to see what a route
expects at a glance, and secondly it allows part of the route to process
before failing with a 422.

Instead, this uses structs to determine what each route should expect.
This means that `actix-web` will never call a handler with an unfinished
payload, allowing it to fail earlier on a request. It also allows
developers to see the arguments for a function more easily based on
their struct definitions.

* Refactor some routes and their paths

Improve the consistency of route paths and their internal handling. Move
some request parameters into the url itself as is common for the others,
updating the frontend and DCL where needed.

* Add a failing test for #269

Add a test that fails on purpose that should test #269. This prevents
this branch from being merged without fixing the issue at some point.

* Remove the last raw `Document` payload

When filtering users, require a JSON payload as with all the other
requests. This is unused on the frontend, so backend changes can be made
without affecting it.

* Improve logging and use stricter typing

Improve the logging within the API server to display slightly less
information and add stricter typing in the payload options for beginning
processing by using the enum itself instead of matching later.

* Add better logging at `debug` and `warn` levels

Improve the logging to be more verbose by using the `debug` level for
general system events and `warn` for when a request fails to reach the
end or a condition is violated, potentially helping developers know why
something occurred.

* Update documentation strings

Make the documentation strings consistent across all routes.

* Add more tests for client-related routes

Add some more tests for the client-related routes, along with
refactoring the `AccessToken` slightly to allow it to be unit tested
more easily.

* Check status codes after protocol requests

When the DCL queries the API server for information about a client, such
as whether they should be authorised, check the response code of the
request for error messages. If the request was not successful, return an
error from the protocol and do not add the client to the nodepool.

Fixes #269.

* Reduce some sleeps and add comments

Reduce some of the sleeps during tests to be using milliseconds instead
of seconds and document the protocol tests slightly better.

* Add a CODEOWNERS file

Add a CODEOWNERS file to allow automatic reviewers depending on the
component changed.

* Respond to expired tokens with 401 Unauthorized

When a client's token has expired, respond with a 401 Unauthorized
instead of a 200 OK along with a new challenge to complete.