Replace interface with Kafka
Adds rdkafka as dependency. Removes interface module.
In the API server replaces forward_to_interface with produce_message
This will add a job_config message to Kafka with the Job_id as a key.
In the DCL with interface_end has been replaced with a loop which
is subscribed to the jobs topic waiting for new jobs.