Shorten again by serializing as bytes
When getting the content of the payload, using `str` as the type will
ensure that it is valid UTF8. Instead, we can just interpret it as a
slice of bytes and allow `serde_json` to fail if this is invalid later.
`rdkafka`'s implementation of `FromBytes` for `[u8]` always returns `Ok`
(it has a unit error type), so the `unwrap` here cannot fail.