Refactor Kafka stream consumption (#291)
* Refactor Kafka stream consumption

Refactor the stream consumption from Kafka within the DCL to reduce
rightward drift. Ignore any Kafka errors for now implicitly, these don't
even include network issues so are unlikely to change anything. Extract
values in `match` statements so that most things happen inline and read
more easily.

* Shorten again by serializing as bytes

When getting the content of the payload, using `str` as the type will
ensure that it is valid UTF8. Instead, we can just interpret it as a
slice of bytes and allow `serde_json` to fail if this is invalid later.
`rdkafka`'s implementation of `FromBytes` for `[u8]` always returns `Ok`
(it has a unit error type), so the `unwrap` here cannot fail.