Replace interface with Kafka (#290)
* Replace interface with Kafka

Adds rdkafka as dependency. Removes interface module.
In the API server replaces forward_to_interface with produce_message
This will add a job_config message to Kafka with the Job_id as a key.
In the DCL with interface_end has been replaced with a loop which
is subscribed to the jobs topic waiting for new jobs.

* Removes the legacy interface test

* Removes interface as member of workspace

* Refactor Kafka stream consumption (#291)

* Refactor Kafka stream consumption

Refactor the stream consumption from Kafka within the DCL to reduce
rightward drift. Ignore any Kafka errors for now implicitly, these don't
even include network issues so are unlikely to change anything. Extract
values in `match` statements so that most things happen inline and read
more easily.

* Shorten again by serializing as bytes

When getting the content of the payload, using `str` as the type will
ensure that it is valid UTF8. Instead, we can just interpret it as a
slice of bytes and allow `serde_json` to fail if this is invalid later.
`rdkafka`'s implementation of `FromBytes` for `[u8]` always returns `Ok`
(it has a unit error type), so the `unwrap` here cannot fail.

Co-authored-by: Alexander Jackson <alexanderjackson@protonmail.com>