Allow models to be unlocked on the frontend
This seems to allow models to be unlocked when using the frontend
instead of sending a raw request through CURL. Previously, Vue would
display error messages about not sending the request in the console and
not actually send a request.

This comes with the caveat that the page does not reload after the
request is sent, but users can refresh the page for now.

Closes #201, #248.