Make changes required for deployment
When deploying the project to the DigitalOcean server, a lot of
hardcoded constants needed to be changed. This includes the name of the
database (which is `sybl-tech`) on the server and MongoDB instance, the
base URL for requests and the binding of the DCL node end socket to
`0.0.0.0` instead of `localhost`.