Update the base URL code to use `VUE_APP` instead
By default, Vue.js will provide `BASE_URL` as the value of `public_path`
in `vue.config.js`. This causes Axios to make requests to the incorrect
path, rendering the website useless. Instead, we can use
`VUE_APP_AXIOS_BASE` as our environment variable, allowing it to be
changed on the server.

This shouldn't change anything for local development, but seems to be
best practice for deployments.