feat: make JWT keys secret in production (#324)
When building a development version of the API server, use the existing
`jwt_key` which is committed to the repository. When building in release
mode for production, use a different file that is not tied to Git. This
causes no existing changes to development workflows, but forces the
server to use a different key.

Closes #209.