tests: fix the multipart upload tests (#332)
Add an asset called `asl.csv` that contains a multipart formatted CSV
file that can be uploaded during tests. This is formatted with DOS line
endings (`\r\n`) as this is what the library expects.

Test requests can then use this file as their payload and set the
`Content-Type` to `form/multipart` to correctly upload the data.