refactor: use `chunks` over a manual approach
Slices have a `chunks` function on them which produces an iterator over
chunks of a given size. This is a lot simpler than the previous approach
and is easier to read. `chunks` will always iterate the entire slice,
producing a chunk of size less or equal to `chunk_size` at the end.

Increase the chunk size to 1000 for better performance, as otherwise we
are storing relatively small documents.