Freddie/gridfs (#333)
* changing frontend to accept multipart

* Begin adding large file support

Currently, users are constrained to approximately 32Kb of data in their
uploaded datasets. Realistically, many users are going to want to
process huge amounts more, so we need to allow larger upload sizes.

The easiest way to do this is through the multipart form specification.
This allows unlimited upload sizes, as long as the web server can
continue to handle it. However, we then need to actually store the files
in a viable way for future processing.

MongoDB advises against having large documents in the database, with the
theoretical boundary being 16MB. While MongoDB has large file support in
the form of GridFS, the current driver does not support it. Instead, we
can make our own GridFS-like models, which contain a `File` and a
`Chunk` for streaming data. This allows us to store large files
efficiently and also process them through streaming later.

Currently, this is very experimental. The API server will happily upload
the data, but no additional processing is done and they are not linked
well to the DCL and other components. Instead, a blank dataset will be
stored where other components expect to find it.

* changing frontend to accept multipart

* working out frontend multipart

* Can stream data, need to change fe to use

* renders larger datasets

* can upload dataset and request training or prediction data

* send analytics server message

* updated predictions endpoint

* added gridfs to prediction storage

* added gridfs to dcl

* added gridfs to analytics server

* fixed test

* commented out tests that don't work

* tests: fix the multipart upload tests (#332)

Add an asset called `asl.csv` that contains a multipart formatted CSV
file that can be uploaded during tests. This is formatted with DOS line
endings (`\r\n`) as this is what the library expects.

Test requests can then use this file as their payload and set the
`Content-Type` to `form/multipart` to correctly upload the data.

* update project status

* render predictions properly

* better prediction cleanup

* more consistent chunking

* load in prediction data

* Can do train and predict upload

* fixed frontend predict rendering

* Changes part 1

* fixing tests for changes made

* more versatile route

* updated tests

* refactor: use `chunks` over a manual approach

Slices have a `chunks` function on them which produces an iterator over
chunks of a given size. This is a lot simpler than the previous approach
and is easier to read. `chunks` will always iterate the entire slice,
producing a chunk of size less or equal to `chunk_size` at the end.

Increase the chunk size to 1000 for better performance, as otherwise we
are storing relatively small documents.

* fixed predictions deletion

Co-authored-by: Alexander Jackson <alexanderjackson@protonmail.com>