feat: use timestamps for computation timeouts
On the frontend side, rename the `timeout` field to
`node_computation_time` as this is more descriptive of what the field
represents. This is then renamed throughout the backend.

When sending the configuration to the client (and converting a
`JobConfiguration` to a `ClientMessage`), include the time that the
message was created. Sending only the cutoff time would be
counterproductive, as clients may not receive the message immediately.
Sending the creation time alongside it allows them to calculate the
timeout.

Providing timestamps is better than timeouts, as clients have a
guaranteed deadline by which to return results that is measured
universally. This means they can return early if needed by checking the
time on each epoch and doing more complex analysis.

Mallus changes will be added to this pull request when both sides have
been approved, however G-Kemp101/mallus#25 has been opened already.