refactor: require both train and predict (#348)
* fix: swap to DOS line endings, add `name` fields

Fix the formatting of the multipart form with multiple uploads to use
DOS line endings (`\r\n` instead of just `\n`) and add `name` fields
which are used by the API server to detect which of the two datasets the
current field is.

* refactor: remove `Option<T>` from datasets

Since datasets should always have both a training and prediction set, it
doesn't make sense for them to be optional in the schema. This change
meant that the previous CSV file needed to be updated, as before it
would not actually upload any data.

* tests: add tests for invalid multipart uploads

Add tests for cases when users manage to only upload a training or
prediction dataset and ensure that they fail.