Move project state to vuex store
Created a new file which contains the state. This store all the projects
and has methods for getting and setting the state. Most API
calls have been moved to this file which update the state accordingly.
The file input is still displayed generated per component but is
currently not working.