Add Vuex state to project (#375)
* Move project state to vuex store

Created a new file which contains the state. This store all the projects
and has methods for getting and setting the state. Most API
calls have been moved to this file which update the state accordingly.
The file input is still displayed generated per component but is
currently not working.

* Change return value of API routes

The routes did not return very useful information so more detailed
responses have been added. `get_user_projecst` now returns the
same information as `get_project` which includes the details and
analysis. The upload data routes now both return the updated project
information.

* Add vuex-persisted state

* Add better authentication and Navigation guards

Adds authentication guards to every route checking if authentication is
needed. If auth is required and user_data is not loaded fetch that.
On dashboard automatically navigate to first project.

* Return Forbidden on unauthorised route

Previously would return Unauthorised which is only if the server does
not know who you are. Changed to forbidden is you cannot
access that resource.

* removed rebase artifacts

* Fixed Mobile Formatting

* Avatar upload process switched to Vuex

* Fixed router name issue

* Renamed Nodes component to Models

* Updated Code Owners

* Alphabetic ordering

* Global particles, stops animation restarting

* Small Screen formatting

* Formatting change

* Updated Welcome to remove extra particle effect

* Reactive unlocking of models

* Fix for job config, advanced option for optionals

* Removed logging

* Included a right margin for the selected project

* Search bar is now a search type and can be cleared

* Deleting dataset dynamicly updates the dashboard

* Start of project analysis componentization

* Made the training tab the active tab for input

* Fixed the deleteData function

* Analysis fully componentized

* Included attribute list

* Standardised data display

* Code clean up

* fix: update tests to pass

* Switched the settings page to use state

* Added model perfomance to the state

* Fix Private Key route name

* Formatting and clearer Variable names

* Enable persistedState

Co-authored-by: Graham Kemp <grahamk206@gmail.com>
Co-authored-by: Alexander Jackson <alexanderjackson@protonmail.com>