fix: show an error on failed login
When users fail to login, show an error below the form to tell them they
had invalid credentials. This was broken as the Axios request would
throw an error on a 401, meaning the later code would not be reached.

Instead, we can simply use a `try catch` statement to handle the error
correctly and show the tooltip.

Fixes #381.