feat: allow clients to regenerate private keys (#385)
* feat: allow clients to regenerate private keys

Add a button in the settings which allows clients to regenerate their
private keys if they have lost them. This will overwrite the public key
in the database with the new one and redirect them to the page that
shows the private keys normally.

Closes #345.

* feat: return 403 on failing to update private keys

Instead of using a `debug_assert` which will not be checked in
production, instead check the number of clients that had their private
keys updated. If this is not 1, it should be 0 and thus a non-client
must have made a request to the route.

* feat: only show private key button for clients

Only show the button to regenerate private keys for client users.

* tests: add tests for new private key route

Add some tests for the new private key generation route to ensure it can
be queried only by clients and not general users.