feat: track more model statistics
Currently the model cards show hardcoded values for the processing time
and credits earned by each one. Instead, we can track these values in
the models themselves, allowing them to be shown on the frontend.

When models are sent predictions, we start a timer and end it after they
have responded. If they fail to respond, no time will be added to their
processing.

When we update the credits for a given user, we also update the total
credits earned for the associated model.

This also fixes some of the finance-related calculations and bugs that
were discovered along the way.

Closes #371.