feat: track more model statistics (#401)
* feat: track more model statistics

Currently the model cards show hardcoded values for the processing time
and credits earned by each one. Instead, we can track these values in
the models themselves, allowing them to be shown on the frontend.

When models are sent predictions, we start a timer and end it after they
have responded. If they fail to respond, no time will be added to their
processing.

When we update the credits for a given user, we also update the total
credits earned for the associated model.

This also fixes some of the finance-related calculations and bugs that
were discovered along the way.

Closes #371.

* chore: requested changes

Update some text on the frontend to better describe the values being
displayed.

* fix: update the tests to pass

Update the reimbursement test to take a model identifier instead of a
user identifier as the function signature has changed slightly.

* docs: update `increment_run_count` docs

Update the `increment_run_count` documentation as it got renamed to
`update_model_statistics` and thus does slightly more than before.