fix: insert datasets before informing analytics (#409)
When uploading data to projects, previously we would message the
analytics server prior to actually uploading the data. This would cause
spurious errors, as sometimes the analytics server would receive the
message after the dataset had been uploaded anyway, allowing it to
continue normally.

However, sometimes it would get the message and the API Server would not
have uploaded the dataset itself yet. This would cause the analytics
server to query for a dataset that does not exist, resulting in an
error.

Instead, we can just make sure we send the message after the dataset
itself has been inserted to the database.