docs: add a guide on the deployed system (#435)
* docs: add a guide on the deployed system

Add a document discussing the deployed system and how it works. This was
requested by @G-Kemp101 and explains the directory structure, continuous
deployment and how to manually make changes.

* docs: add sections requested in review

Add sections on `systemd` which handles `kafka` and `mongod`, the local
database instance itself and the persistent models running on the
system.