fix: deadlock when failing to build clusters
Firstly, reset models to not be in use if a cluster fails to be built
when flooding the network. This was causing an issue where models would
not be considered for jobs if they accepted a previous one that failed
to be run.

For example, if a job with 2 models is requested and one accepts and the
other rejects, the accepting model would be permanently stuck and would
not receive later jobs. This is solved by iterating through and
unlocking models if the cluster fails to be built.

This causes a deadlock in the `NodePool::end` function, as the
`build_cluster` function still has a lock on the `info` field of the
pool.

This can be fixed by calling `drop` on the `RwLockWriteGuard` held by
`build_cluster` as soon as the loop finishes, allowing it to be accessed
later on if the cluster has failed to be built.