Fix active increment and decrement
The `NodePool::active` var was incremented before it was decremented if
the cluster failed to build. The decrement was moved to the match on the
job config.

active was also decremented again in dcl protocol if a job failed
during processing. Added a check on `update_node_alive` to only dec
if the node is not `using`