feat: distribute the control layer
Begin the process of distributing the control layer. Running with `cargo
run -- control` will begin the control node and otherwise will start a
regular edge node as before. Changing to this approach means that you
need to run both a control node and an edge node, so this is mostly
experimental.

The control node performs 2 functions. Firstly, it will allow edge nodes
to connect and begin heartbeating, allowing themselves to be registered
as a potential connection point for clients. Secondly, clients from
`mallus` can make requests for a port to connect to, which will return a
port that an active edge node is connected on.

Clients are thus expected to first query the control node on port 7000
(usually) and then connect normally to the port provided as a response.
Communications from here are as normal and have not changed otherwise.

Edge nodes now need to heartbeat with the control node. This is done
through the `health` module and runs alongside the health checking of
client nodes.

Some additional messages have been added specific to the `control`
portion of the system. This includes messages for heartbeating as
normal, but also the edge node registration and port requests, alongside
the port responses if one exists.