Fix/dcl bugs (#462)
* Change when nodepool active increases

The nodepool active counter used to increase on every new connection.
If a node disconnected and reconnected during a heartbeat
 the active count would increase but the active nodes would not.
 Added a check to only increase active count if new node added is not
 already marked as active

* Fix active increment and decrement

The `NodePool::active` var was incremented before it was decremented if
the cluster failed to build. The decrement was moved to the match on the
job config.

active was also decremented again in dcl protocol if a job failed
during processing. Added a check on `update_node_alive` to only dec
if the node is not `using`